% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpreflinux.R
\name{preferencias}
\alias{preferencias}
\title{Modelo de Preferencias StatConsult}
\usage{
preferencias(N, nbeta, nmarcas, nclase, clases, X, y, missing, niter = 15000,
  nburn = 100, nstep = 10, nu = NULL, S = NULL)
}
\arguments{
\item{N}{Individuos}

\item{nbeta}{Cantidad de coeficientes}

\item{nmarcas}{Cantidad de objetos a preferir}

\item{nclase}{Cantidad de posibles clases de coeficientes (nclases=N indica uno por individuo)}

\item{clases}{Mapa entre individuos y clases}

\item{X}{Covariables que explican la preferencia}

\item{y}{Vector con las preferidas (1...nmarcas, 0 para ninguna, NA para valor perdido)}

\item{missing}{Indicador con cuales observaciones faltan}

\item{niter}{Iteraciones del MCMC}

\item{nburn}{Iteraciones a quemar por el MCMC}

\item{nstep}{Pasos de iteraciones para tomar una muestra}

\item{nu}{Parametros de la previa}

\item{S}{Parametros de la previa}
}
\value{
La funcion devuelve sus resultados en unos archivos de texto
}
\description{
Modela la preferencia hacia una marca en particular
versus otras marcas en funcion de las evaluaciones dadas por cada individuo
para las marcas (no necesariamente evaluando las mismas marcas)

logit(p_{i,M_{ij}}) = beta_i X_{i,M_{ij}}
donde p es la probabilidad de preferir la marca particular contra la marca
M_{ij} evaluada por el individuo i. beta_i es el vector de coeficientes,
modelados a priori independientemente como una gamma(nu, lambda_k) para
cada atributo k (un coeficiente por atributo). X contiene la diferencia
entre las evaluaciones de la marca particular y la marca M_{ij} para
cada una de los K atributos. lambda a su vez es modelado como una
iid igamma (alpha, lambda0)

Modela la preferencia hacia una marca en particular
versus otras marcas en funcion de las evaluaciones dadas por cada individuo
para las marcas (no necesariamente evaluando las mismas marcas)
}
\examples{
## buscar los ejemplos de los proyectos de modelos de marca

}

